requirements.txt
discord
python-dotenv
pydexscreener